fix: update vector constructor

+++ Primitives/IObject.hpp
@@ -22,8 +22,8 @@ namespace objects {
         virtual void setPosition(raytracer::Vector3<double> pos) = 0;
         [[nodiscard]] virtual raytracer::Vector3<double> getPosition() = 0;
 
-        virtual void setColor(raytracer::Vector3<int> rgb) = 0;
-        virtual raytracer::Vector3<int> getColor() const = 0;
+        virtual void setColor(raytracer::Vector3<double> rgb) = 0;
+        virtual raytracer::Vector3<double> getColor() const = 0;
 
         virtual std::string getType() const = 0;
     };
